% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_MSSI.R
\name{calculate_MSSI}
\alias{calculate_MSSI}
\title{Function to calculate the Multiscale Straightness Index}
\usage{
calculate_MSSI(data, uniqueID = "traj", time = "frame", window_size,
  granulosity)
}
\arguments{
\item{data}{A dataframe containing the X and Y coordinates}

\item{uniqueID}{The unique identifier for each trajectory in the dataframe}

\item{time}{The column containing the information on the time for each fix}

\item{window_size}{A numeric value or vector specifying the window sizes over which the MSSI should be calculated}

\item{granulosity}{A numeric value or vector specifying the temporal resolution of the data before the MSSI is calculated}
}
\value{
A dataframe containing the MSSI for each fix in addition to the input data
}
\description{
Takes a dataframe with the X and Y coordinates of a trajectory and calculates the MSSI for different window sizes and
granulosities.
}
\references{
C.M. Postlethwaite, P. Brown, und T.E. Dennis: A New Multi-Scale Measure for Analysing Animal Movement Data.
Journal of Theoretical Biology 317 (2013): 175-185. doi:10.1016/j.jtbi.2012.10.007.
}

