% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSSI.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Movement trajectories of aquatic microbes}
\format{A data frame with example trajectories of aquatic microbes}
\usage{
data(example_data)
}
\description{
Movement trajectories of aquatic microbes
}
\keyword{datasets}

